+++ R/utils_functional.R
@@ -1,3 +1,8 @@
+fltr <- function (x, f, ...) {
+  index <- map_lgl(x, f, ...)
+  x[index]
+}
+
 map <- function(x, f, ...) {
   lapply(X = x, FUN = as_function(f), ...)
 }
@@ -8,6 +13,10 @@ map_chr <- function(x, f, ...) {
 
 map_int <- function(x, f, ...) {
   vapply(X = x, FUN = as_function(f), FUN.VALUE = integer(1), ...)
+}
+
+map_lgl <- function (x, f, ...) {
+  vapply(X = x, FUN = as_function(f), FUN.VALUE = logical(1), ...)
 }
 
 set_names <- function(x, names) {